+++ arch/arm/core/cortex_a_r/CMakeLists.txt
@@ -16,6 +16,7 @@ zephyr_library_sources(
   swap_helper.S
   irq_manage.c
   prep_c.c
+  thread.c
   )
 
 zephyr_library_sources_ifdef(CONFIG_GEN_SW_ISR_TABLE isr_wrapper.S)
